+++ docs/CHANGELOG.md
@@ -25,6 +25,7 @@ Changelog
 - Enh #107: Emoji search keywords are not translatable at the moment
 - Enh #82: Simply Tt menu items
 - Fix #97: [Bug] Images duplicate on drag and drop
+- Enh #43: Make menu accessible
 
 
 1.1.3 (Unreleased)